requirements_credentials.txt
pywin32; platform_system == "Windows"             # Another one for windows interface, used for testing username and password
python-pam; platform_system != "Windows"             # For linux or macos, to test passwords
